Fix to drop Exception Levels
Lots of changes necessary make this commit work. Mostly weird behavior
with QEMU.

Changed the `-kernel` target of qemu to the IMG file instead of the elf
file. No idea why but Qemu does things differently if we run off the
elf or the img file. When starting from the elf file, we start at EL3
and dropping EL's seem to be problematic. Have not retested it after
applying the current fix.

Added the `-d int` argument to have Qemu output interrupts. Handy.

Added a bunch of breakpoints in gdbinit to pop on every function during
debugging.

Added a `continue` statement to gdbinit because due to running from the
.img, the kernel is started as not running even when connecting from
GDB.

Lots of new code in boot.s to drop from whatever EL we are on down to
EL1. Not very clean assembly code but it is currently working.