Fixed error vectors
Dumb mistake. By following the tutorial post, the person only
coded the first of four blocks needed per VBAR register. Only the
first section is needed if we are staying at the same exception level
but since this example is going from EL1 to EL2 we need the third
section.

This fixes it by copying the first section 4 times to always do the
same thing. I believe it should be adapted over time but for now this
is working.