Switch back to Raspi3b
When using the virt board it is hard to emulate specific additional
devices. It basically tries to emulate a Linux PC. Stuff like a UART
is not present.

I got inspired by some code online that ran a specific machine + a
specific CPU. This gave me the idea of running a Raspi3 with an A76.

Since devices should be fairly similar between RPis this should do for
now.